requirements.txt
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
filelock==3.16.1
fsspec==2024.9.0
idna==3.10
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
openai-whisper==20240930
PyAudio==0.2.14
pycparser==2.22
regex==2024.9.11
requests==2.32.3
scipy==1.14.1
setuptools==75.1.0
sounddevice==0.5.0
SpeechRecognition==3.10.4
sympy==1.13.3
tiktoken==0.8.0
torch==2.2.2
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
